+++ vite.config.ts
@@ -136,10 +136,5 @@ export default defineConfig({
                 }
             }
         }
-    },
-    test: {
-        deps: {
-            inline: ['arisitd-ds']
-        }
     }
 });